pyproject.toml
[project]
name = "gobby"
version = "0.1.1"
description = "Gobby is a local daemon that unifies Claude Code, Gemini CLI, and Codex through a hook interface for session tracking, and provides an MCP proxy with progressive tool discovery for efficient access to downstream servers."
requires-python = ">=3.11"
dependencies = [
    "click>=8.1.0",
    "fastapi>=0.115.0",
    "fastmcp>=0.2.0",
    "httpx>=0.27.0",
    "litellm>=1.0.0",
    "psutil>=6.1.0",
    "py-machineid>=0.6.0",
    "pydantic>=2.9.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=15.0",
    "pyyaml>=6.0.3",
    "jinja2>=3.1.0",
    "claude-agent-sdk>=0.1.5",
]

[project.scripts]
gobby = "gobby.cli:cli"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
gobby = [
    "install/claude/**/*",
    "install/gemini/**/*",
    "install/codex/**/*",
    "install/antigravity/**/*",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--cov=gobby",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=80",
    "-v",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "e2e: marks tests as end-to-end tests",
]

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = [".archive"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "TID252", # ban relative imports from parent packages (from .. import)
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py

[tool.mypy]
python_version = "3.11"
plugins = ["pydantic.mypy"]
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = false
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[dependency-groups]
dev = [
    "mypy>=1.8.0",
    "pre-commit>=4.0.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "pytest-httpx>=0.30.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.8.0",
    "types-PyYAML>=6.0.12.12",
    "pyrefly>=0.1.0",
]
